pyproject.toml
[project]
name = "hikeplanner"
version = "0.1.0"
description = "HikePlanner"
readme = "README.md"
requires-python = "==3.13.7"
dependencies = [
  "flask==3.1.0",
  "flask-cors==5.0.0",
  "gpxpy==1.6.2",
  "azure-storage-blob==12.24.0",
  "pandas==2.2.3",
  "seaborn==0.13.2",
  "pymongo==4.10.1",
  "python-dotenv==1.2.1",
  "argparse==1.4.0",
  "scikit-learn==1.5.2",
]

[dependency-groups]
dev = []
